REFACTOR: move routes folder to src
